Refactor LoginForm, SignupForm, and LoginSignup components to import Login-SignupStyle.js instead of Login-Singup.js.
