+++ client/src/pages/LoginSignup.jsx
@@ -13,11 +13,20 @@ const MainContainer = styled.div`
   background-color: #f1f1f1;
 `;
 
-const Image = styled.img`  
-  height: 80%;
+const Image = styled.img`
+  height: 70%;
   object-fit: cover;
   margin: 0 auto;
 `
+
+const Slogan = styled.h2`
+  font-size: 2rem;
+  font-weight: 400;
+  color: #555;
+  text-align: center; // Center the slogan text under the image
+  margin: 0 auto; // Adjust the margin as needed
+  margin-bottom: 20px;
+`;
 
 export function LoginSignup({setIsLoggedIn, mode}) {
   const isSignUp = mode === 'signup';
@@ -25,11 +34,14 @@ export function LoginSignup({setIsLoggedIn, mode}) {
   return (
     <>
       <MainContainer>
+        {/*<div>*/}
         <Image src={logo}/>
+        {/*<Slogan>"Step into Style:<br/> Where Every Footprint Tells a Story"</Slogan>*/}
+        {/*</div>*/}
         {isSignUp ? (
-          <SignupForm setIsLoggedIn={setIsLoggedIn} />
+          <SignupForm setIsLoggedIn={setIsLoggedIn}/>
         ) : (
-          <LoginForm setIsLoggedIn={setIsLoggedIn} />
+          <LoginForm setIsLoggedIn={setIsLoggedIn}/>
         )}
       </MainContainer>
     </>
